move checkout form into a separate component
trigger the embeddable checkout form when user clicks on `upgrade to
researcher`. update to import stripe packages for the component.
